Fixed mdt clearing

+++ resources/vrp/base.lua
@@ -669,17 +669,30 @@ Citizen.CreateThread(function()
 	end
 end)
 
+local osHour = nil
+local timeout = 30
+
 Citizen.CreateThread(function()
+	print("MDT Debug - Cleanup wait started")
 	Citizen.Wait(10000)
+	print("MDT Debug - Cleanup thread started")
 	local osTime = os.date("*t")
 	if osTime ~= nil then
-		local osHour = osTime.hour+1
-		if (osHour) == 2 or (osHour) == 8 or (osHour) == 14 or (osHour) == 20 then
-			Citizen.Wait(60000*30) -- wait 30 minutes
-			if osHour == 2 then
-				MySQL.Async.execute('DELETE FROM gta5_gamemode_essential.vrp_mdt WHERE HOUR(CAST(dateInserted AS TIME)) < @insertedHour or HOUR(CAST(dateInserted AS TIME)) > 9', {insertedHour = osHour}, function(rowsChanged)	end)
-			else
-				MySQL.Async.execute('DELETE FROM gta5_gamemode_essential.vrp_mdt WHERE HOUR(CAST(dateInserted AS TIME)) < @insertedHour', {insertedHour = osHour}, function(rowsChanged)	end)
+		osHour = osTime.hour
+		print("MDT Debug - OS Hour = "..osHour)
+		if osHour ~= nil and ((osHour) == 2 or (osHour) == 8 or (osHour) == 14 or (osHour) == 20) then
+			print("MDT Debug - Delayed Loop Started")
+			while timeout > 0 do
+				Citizen.Wait(60000)
+				timeout = timeout - 1
+				if timeout <= 0 then
+					print("MDT Debug - Cleanup Iniated")
+					if osHour == 2 then
+						MySQL.Async.execute('DELETE FROM gta5_gamemode_essential.vrp_mdt WHERE HOUR(CAST(dateInserted AS TIME)) < @insertedHour or HOUR(CAST(dateInserted AS TIME)) > 3', {insertedHour = osHour}, function(rowsChanged)	end)
+					else
+						MySQL.Async.execute('DELETE FROM gta5_gamemode_essential.vrp_mdt WHERE HOUR(CAST(dateInserted AS TIME)) < @insertedHour', {insertedHour = osHour}, function(rowsChanged)	end)
+					end
+				end
 			end
 		end
 	end
